--- src/main/java/com/example/ogiyo/ceoReview/dto/request/SaveCeoReviewRequestDto.java
@@ -1,4 +1,0 @@
-package com.example.ogiyo.ceoReview.dto.request;
-
-public class SaveCeoReviewRequestDto {
-}
