--- src/main/java/com/example/ogiyo/ceoReview/repository/CeoReviewRepository.java
@@ -1,8 +1,0 @@
-package com.example.ogiyo.ceoReview.repository;
-
-import com.example.ogiyo.ceoReview.entity.CeoReview;
-import com.example.ogiyo.review.entity.Review;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CeoReviewRepository extends JpaRepository<CeoReview, Long> {
-}
