--- src/main/java/com/example/ogiyo/ceoReview/service/CeoReviewService.java
@@ -1,4 +1,0 @@
-package com.example.ogiyo.ceoReview.service;
-
-public class CeoReviewService {
-}
