+++ src/main/java/com/example/ogiyo/domain/ceoReview/dto/request/SaveCeoReviewRequestDto.java
@@ -1,0 +1,4 @@
+package com.example.ogiyo.domain.ceoReview.dto.request;
+
+public class SaveCeoReviewRequestDto {
+}
