+++ src/main/java/com/example/ogiyo/domain/ceoReview/repository/CeoReviewRepository.java
@@ -1,0 +1,7 @@
+package com.example.ogiyo.domain.ceoReview.repository;
+
+import com.example.ogiyo.domain.ceoReview.entity.CeoReview;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CeoReviewRepository extends JpaRepository<CeoReview, Long> {
+}
