+++ src/main/java/com/example/ogiyo/domain/ceoReview/service/CeoReviewService.java
@@ -1,0 +1,4 @@
+package com.example.ogiyo.domain.ceoReview.service;
+
+public class CeoReviewService {
+}
