+++ src/main/java/com/example/ogiyo/domain/review/repository/ReviewRepository.java
@@ -1,6 +1,6 @@
-package com.example.ogiyo.review.repository;
+package com.example.ogiyo.domain.review.repository;
 
-import com.example.ogiyo.review.entity.Review;
+import com.example.ogiyo.domain.review.entity.Review;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
